cloud/requirements.txt
functions-framework==3.*
cloudevents==1.*
google-cloud-pubsub==2.*
